Move Markdown component to its own directory
